feat(create_user.py): add use case for creating a user
The `CreateUser` use case is responsible for handling the creation of a user. It takes a `CreateUserDto` object as input and returns a `User` object.

The use case initializes a `CreateUserMapper` to map the `CreateUserDto` to a `User` object. The `handle` method then uses the mapper to map the input `user` to a `User` object. The `password` field is then removed from the `User` object before returning it.